# Optimize cell_to_children O(n^2) to O(n)